[Fixes #11992] Transfer users API under people app
