+++ geonode/base/api/urls.py
@@ -20,7 +20,6 @@ from geonode.api.urls import router
 
 from . import views
 
-router.register(r"users", views.UserViewSet, "users")
 router.register(r"groups", views.GroupViewSet, "group-profiles")
 router.register(r"resources", views.ResourceBaseViewSet, "base-resources")
 router.register(r"owners", views.OwnerViewSet, "owners")
