+++ geonode/people/views.py
@@ -38,6 +38,24 @@ from geonode.base.views import user_and_group_permission
 
 from dal import autocomplete
 
+from dynamic_rest.filters import DynamicFilterBackend, DynamicSortingFilter
+from oauth2_provider.contrib.rest_framework import OAuth2Authentication
+from drf_spectacular.utils import extend_schema
+from dynamic_rest.viewsets import DynamicModelViewSet
+from rest_framework.response import Response
+from rest_framework.decorators import action
+from rest_framework.permissions import IsAuthenticated
+from geonode.base.models import ResourceBase
+from geonode.base.api.filters import DynamicSearchFilter
+from geonode.groups.models import GroupProfile, GroupMember
+from geonode.base.api.permissions import IsSelfOrAdminOrReadOnly
+from geonode.base.api.serializers import UserSerializer, GroupProfileSerializer, ResourceBaseSerializer
+from geonode.base.api.pagination import GeoNodeApiPagination
+
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication
+from geonode.security.utils import get_visible_resources
+from guardian.shortcuts import get_objects_for_user
+
 
 class SetUserLayerPermission(View):
     def get(self, request):
@@ -142,6 +160,71 @@ def forgot_username(request):
                 message = _("No user could be found with that email address.")
 
     return render(request, "people/forgot_username_form.html", context={"message": message, "form": username_form})
+
+
+class UserViewSet(DynamicModelViewSet):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+
+    authentication_classes = [SessionAuthentication, BasicAuthentication, OAuth2Authentication]
+    permission_classes = [
+        IsAuthenticated,
+        IsSelfOrAdminOrReadOnly,
+    ]
+    filter_backends = [DynamicFilterBackend, DynamicSortingFilter, DynamicSearchFilter]
+    serializer_class = UserSerializer
+    pagination_class = GeoNodeApiPagination
+
+    def get_queryset(self):
+        """
+        Filters and sorts users.
+        """
+        if self.request and self.request.user:
+            queryset = get_available_users(self.request.user)
+        else:
+            queryset = get_user_model().objects.all()
+
+        # Set up eager loading to avoid N+1 selects
+        queryset = self.get_serializer_class().setup_eager_loading(queryset)
+        return queryset.order_by("username")
+
+    @extend_schema(
+        methods=["get"],
+        responses={200: ResourceBaseSerializer(many=True)},
+        description="API endpoint allowing to retrieve the Resources visible to the user.",
+    )
+    @action(detail=True, methods=["get"])
+    def resources(self, request, pk=None):
+        user = self.get_object()
+        permitted = get_objects_for_user(user, "base.view_resourcebase")
+        qs = ResourceBase.objects.all().filter(id__in=permitted).order_by("title")
+
+        resources = get_visible_resources(
+            qs,
+            user,
+            admin_approval_required=settings.ADMIN_MODERATE_UPLOADS,
+            unpublished_not_visible=settings.RESOURCE_PUBLISHING,
+            private_groups_not_visibile=settings.GROUP_PRIVATE_RESOURCES,
+        )
+
+        paginator = GeoNodeApiPagination()
+        paginator.page_size = request.GET.get("page_size", 10)
+        result_page = paginator.paginate_queryset(resources, request)
+        serializer = ResourceBaseSerializer(result_page, embed=True, many=True, context={"request": request})
+        return paginator.get_paginated_response({"resources": serializer.data})
+
+    @extend_schema(
+        methods=["get"],
+        responses={200: GroupProfileSerializer(many=True)},
+        description="API endpoint allowing to retrieve the Groups the user is member of.",
+    )
+    @action(detail=True, methods=["get"])
+    def groups(self, request, pk=None):
+        user = self.get_object()
+        qs_ids = GroupMember.objects.filter(user=user).values_list("group", flat=True)
+        groups = GroupProfile.objects.filter(id__in=qs_ids)
+        return Response(GroupProfileSerializer(embed=True, many=True).to_representation(groups))
 
 
 class ProfileAutocomplete(autocomplete.Select2QuerySetView):
